Remove the concept of manager
This was an experiment and ended up being a bad idea. It added
complexity and constraints on the user's configuration whitout
having substantial benefits. Managers can be implemented as regular
modules with more flexibility and control over the loading order.

In particular, some "managers" like prezto require to be loaded last,
and not first as the concept was enforcing. Additionally, some plugin
managers configuraton might also want to register for more than a single
runcom. Both of these invalidate the original assumption that gave birth to
managers, rendering them useless and harmful to the global design.